```text feat: Improve SQL script for ranking country origins of bands by number of fans
The SQL script `2-fans.sql` has been modified to improve the ranking of country origins of bands based on the number of fans. The script now uses the `SUM` function instead of `COUNT` to calculate the total number of fans. This change will distribute the load better and provide more accurate results.