feat: Add stored procedure to compute and store average score for a student
This commit adds a new SQL script `7-average_score.sql` that creates a stored procedure `ComputeAverageScoreForUser` to compute and store the average score for a student. The procedure takes a parameter `user_id` (integer) and calculates the average score by selecting the average of the `score` column from the `corrections` table where the `user_id` matches the provided parameter. The calculated average score is then updated in the `users` table for the corresponding `user_id`. This script can be executed on any database.

The commit also includes the necessary delimiter statements to define the start and end of the procedure.